wsd: move STATE_ENUM to common/StateEnum.hpp
With some minor modifications.
